# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SmartNotify AI - WhatsApp Notification Router API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: messages
    description: Message prediction operations
  - name: dashboard
    description: Dashboard and analytics
  - name: upload
    description: Dataset upload
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List all messages with predictions
      description: Returns all messages from the loaded dataset with their AI predictions
      parameters:
        - name: action
          in: query
          schema:
            type: string
          description: Filter by action (notify, digest, mute)
        - name: messageType
          in: query
          schema:
            type: string
          description: Filter by message type
        - name: conversationType
          in: query
          schema:
            type: string
          description: Filter by conversation type (personal, group, business)
        - name: search
          in: query
          schema:
            type: string
          description: Search in message text
      responses:
        "200":
          description: List of messages with predictions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessagePrediction"

  /messages/{messageId}:
    get:
      operationId: getMessage
      tags: [messages]
      summary: Get a single message with full prediction detail
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Message with prediction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePrediction"
        "404":
          description: Message not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /messages/predict:
    post:
      operationId: predictMessage
      tags: [messages]
      summary: Predict routing for a single message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredictInput"
      responses:
        "200":
          description: Prediction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prediction"

  /messages/predict-all:
    post:
      operationId: predictAll
      tags: [messages]
      summary: Run predictions on all loaded messages
      responses:
        "200":
          description: Bulk prediction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkPredictResult"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard summary statistics
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /analytics:
    get:
      operationId: getAnalytics
      tags: [dashboard]
      summary: Get analytics data for charts
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsData"

  /upload:
    post:
      operationId: uploadDataset
      tags: [upload]
      summary: Upload and process dataset CSV files
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadInput"
      responses:
        "200":
          description: Upload and processing result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"

  /export:
    get:
      operationId: exportPredictions
      tags: [messages]
      summary: Export predictions as CSV download URL
      responses:
        "200":
          description: Export result with download URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    Message:
      type: object
      required:
        - messageId
        - userId
        - conversationType
        - createdAt
        - forwardedCount
      properties:
        messageId:
          type: string
        userId:
          type: string
        conversationType:
          type: string
          enum: [personal, group, business]
        groupId:
          type: ["string", "null"]
        businessId:
          type: ["string", "null"]
        senderUserId:
          type: ["string", "null"]
        createdAt:
          type: string
        messageText:
          type: ["string", "null"]
        mediaType:
          type: ["string", "null"]
          enum: ["image", "voice", "null", null]
        mediaId:
          type: ["string", "null"]
        forwardedCount:
          type: number

    Prediction:
      type: object
      required:
        - action
        - messageType
        - reason
        - confidence
        - evidenceMessageIds
        - scamProbability
        - spamProbability
        - urgencyScore
        - businessTrustScore
      properties:
        action:
          type: string
          enum: [notify, digest, mute]
        messageType:
          type: string
          enum: [personal, urgent, event, payment, business_update, promotion, greeting, forward, spam, scam, unknown]
        reason:
          type: string
        confidence:
          type: number
        evidenceMessageIds:
          type: string
        scamProbability:
          type: number
        spamProbability:
          type: number
        urgencyScore:
          type: number
        businessTrustScore:
          type: number
        personalizationFactors:
          type: array
          items:
            type: string

    MessagePrediction:
      type: object
      required:
        - message
        - prediction
      properties:
        message:
          $ref: "#/components/schemas/Message"
        prediction:
          $ref: "#/components/schemas/Prediction"

    DashboardStats:
      type: object
      required:
        - totalMessages
        - notifyCount
        - digestCount
        - muteCount
        - scamCount
        - spamCount
        - urgentCount
        - averageConfidence
        - messagesByType
        - recentMessages
      properties:
        totalMessages:
          type: number
        notifyCount:
          type: number
        digestCount:
          type: number
        muteCount:
          type: number
        scamCount:
          type: number
        spamCount:
          type: number
        urgentCount:
          type: number
        averageConfidence:
          type: number
        messagesByType:
          type: array
          items:
            $ref: "#/components/schemas/CountByType"
        recentMessages:
          type: array
          items:
            $ref: "#/components/schemas/MessagePrediction"

    CountByType:
      type: object
      required: [type, count]
      properties:
        type:
          type: string
        count:
          type: number

    AnalyticsData:
      type: object
      required:
        - actionDistribution
        - messageTypeDistribution
        - conversationTypeBreakdown
        - confidenceDistribution
        - dailyTrend
        - topSenders
      properties:
        actionDistribution:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        messageTypeDistribution:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        conversationTypeBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        confidenceDistribution:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        dailyTrend:
          type: array
          items:
            $ref: "#/components/schemas/DailyTrendPoint"
        topSenders:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"

    ChartDataPoint:
      type: object
      required: [name, value]
      properties:
        name:
          type: string
        value:
          type: number
        color:
          type: ["string", "null"]

    DailyTrendPoint:
      type: object
      required: [date, notify, digest, mute]
      properties:
        date:
          type: string
        notify:
          type: number
        digest:
          type: number
        mute:
          type: number

    PredictInput:
      type: object
      required:
        - messageText
        - conversationType
        - forwardedCount
      properties:
        messageText:
          type: string
        conversationType:
          type: string
          enum: [personal, group, business]
        groupId:
          type: ["string", "null"]
        businessId:
          type: ["string", "null"]
        senderUserId:
          type: ["string", "null"]
        mediaType:
          type: ["string", "null"]
        forwardedCount:
          type: number

    BulkPredictResult:
      type: object
      required: [processed, notifyCount, digestCount, muteCount]
      properties:
        processed:
          type: number
        notifyCount:
          type: number
        digestCount:
          type: number
        muteCount:
          type: number

    UploadInput:
      type: object
      required: [datasetName]
      properties:
        datasetName:
          type: string
        csvContent:
          type: ["string", "null"]

    UploadResult:
      type: object
      required: [success, message, messagesLoaded]
      properties:
        success:
          type: boolean
        message:
          type: string
        messagesLoaded:
          type: number

    ExportResult:
      type: object
      required: [downloadUrl, totalRows]
      properties:
        downloadUrl:
          type: string
        totalRows:
          type: number
        csvContent:
          type: string